Moved mixed strategies calculation to parent game class
