+++ src/app/classes/Game.class.ts
@@ -1,5 +1,5 @@
 import { BehaviorSubject, Observable } from 'rxjs';
-import { Matrix } from '.';
+import { Matrix, MixedStrategy } from '.';
 import { Results } from '../services/game-service.interface';
 import Player from './Player.class';
 import Round from './Round.class';
@@ -124,4 +124,29 @@ export default abstract class Game {
   }
 
   abstract getGameResults(): Results;
+
+  protected calculateMixedStrategies(): MixedStrategy[] {
+    const mixedStrategies: MixedStrategy[] = [];
+
+    for (const player of this.players) {
+      const playedStrategies = this.rounds.map(round => round.playedStrategies[player.playerNumber].strategy);
+      const playerStrategies = this.matrix.playersStrategies[player.playerNumber];
+
+      const strategyCounter = playerStrategies.reduce((counter, strategy) => { counter[strategy] = 0; return counter; }, {});
+
+      playedStrategies.forEach(strategy => strategyCounter[strategy]++);
+
+      const mixedStrategy = new MixedStrategy(
+        player,
+        playerStrategies.map(strategy => ({ 
+          strategy,
+          probability: strategyCounter[strategy] / this.rounds.length
+        }))
+      );
+
+      mixedStrategies.push(mixedStrategy);
+    }
+
+    return mixedStrategies;
+  }
 }
